>update seed data

+++ Models/NiveauScolaire.cs
@@ -8,10 +8,10 @@ namespace Models
         public string Nom { get; set; }
         public string NomAr { get; set; }
         public int IdCycle { get; set; }
-        public int CoursLigneGroupe { get; set; }
-        public int CoursLigneIndividuel { get; set; }
-        public int CoursDomicileGroupe { get; set; }
-        public int CoursDomicileIndividuel { get; set; }
+        public double CoursLigneGroupe { get; set; }
+        public double CoursLigneIndividuel { get; set; }
+        public double CoursDomicileGroupe { get; set; }
+        public double CoursDomicileIndividuel { get; set; }
         public virtual ICollection<Branche> Branches { get; set; }
         public virtual ICollection<Cours> Courses { get; set; }
